[Entity Store] Use START_WITH instead of like
